services/api-gateway/auth-sidecar/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-sidecar"
version = "0.1.0"
description = "SmartCourse Auth Sidecar — JWT verification for Nginx"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
